Move µEd OpenAPI spec into runtime/schema
Relocates the spec from api/ into runtime/schema/ alongside the existing
JSON schema files, and renames it to mued_v0.1.0.yml to make the version
explicit. Removes the api/ package; embed is now owned by runtime/schema.
