+++ runtime/schema/openapi.go
@@ -1,6 +1,6 @@
-package api
+package schema
 
 import _ "embed"
 
-//go:embed openapi.yml
+//go:embed mued_v0.1.0.yml
 var OpenAPISpec []byte
